db/seg: caller-managed SAIS scratch buffer; drop legacy patricia MatchFinder
Sais takes a caller-provided *[]int32 scratch buffer instead of an internal
sync.Pool. Callers reuse one buffer per worker: extractPatternsInSuperstrings
via a package-level pool, MatchFinder3 via a struct field. Also removes the
unused pointer-based patricia trie (MatchFinder/MatchFinder2), keeping the
flat-tree implementation.
