+++ db/seg/parallel_compress.go
@@ -957,6 +957,8 @@ func copyN(r io.Reader, w io.Writer, uncoveredCount int, buf []byte) error {
 	return nil
 }
 
+var saisBufPool = sync.Pool{New: func() any { return new([]int32) }}
+
 // extractPatternsInSuperstrings is the worker that processes one superstring and puts results
 // into the collector, using lock to mutual exclusion. At the end (when the input channel is closed),
 // it notifies the waitgroup before exiting, so that the caller known when all work is done
@@ -966,7 +968,9 @@ func extractPatternsInSuperstrings(ctx context.Context, superstringCh chan []byt
 	defer completion.Done()
 	dictVal := make([]byte, 8)
 	dictKey := make([]byte, maxPatternLen)
-	var lcp, sa, inv, saisBuf []int32
+	saisBuf := saisBufPool.Get().(*[]int32)
+	defer saisBufPool.Put(saisBuf)
+	var lcp, sa, inv []int32
 	for {
 		var (
 			superstring []byte
@@ -988,7 +992,7 @@ func extractPatternsInSuperstrings(ctx context.Context, superstringCh chan []byt
 		}
 		//log.Info("Superstring", "len", len(superstring))
 		//start := time.Now()
-		if err := sais.Sais(superstring, sa, &saisBuf); err != nil {
+		if err := sais.Sais(superstring, sa, saisBuf); err != nil {
 			panic(err)
 		}
 		//log.Info("Suffix array built", "in", time.Since(start))
